refactor: changed to compound components, types are defined in another file
