+++ components/Quote/Quotes.tsx
@@ -4,7 +4,7 @@ import QuoteSlider from './QuoteSlider';
 
 const Quotes = () => {
   return (
-    <div className="my-16 mx-4 flex flex-col items-center justify-center md:mx-28 md:my-24 md:flex-row">
+    <div className="mx-4 my-16 flex flex-col items-center justify-center md:mx-28 md:my-24 md:flex-row">
       <QuoteSlider />
     </div>
   );
