+++ data/quotes.ts
@@ -1,12 +1,4 @@
-export interface Quote {
-  id: number;
-  content: string;
-  author: {
-    name: string;
-    title: string;
-    avatar: string;
-  };
-}
+import type { Quote } from 'types';
 
 const Quotes: Quote[] = [
   {
