+++ types/index.ts
@@ -1,0 +1,9 @@
+export interface Quote {
+  id: number;
+  content: string;
+  author: {
+    name: string;
+    title: string;
+    avatar: string;
+  };
+}
